Bump freq from int to long
There are cases where freq can have very large values (30-40 million
or so).  The loop in adjmeans() can overflow because the multiplication
happens with int types.

This bug made the algorithm not converge for some data sets.

+++ colors.c
@@ -17,7 +17,7 @@ struct point {
 	int x;
 	int y;
 	int z;
-	int freq;
+	long freq;
 	struct cluster *c;
 	TAILQ_ENTRY(point) e;
 	RB_ENTRY(point) rb_e;
